Added a dialog to confirm deletion of task.

+++ apps/mobile/lib/main.dart
@@ -108,9 +108,40 @@ class _AppState extends State<App> {
                           ? IconButton(
                               icon: const Icon(Icons.delete),
                               color: Colors.red,
-                              onPressed: () {
-                                task.delete();
-                                setState(() {});
+                              onPressed: () async {
+                                final confirmed = await showDialog<bool>(
+                                  context: context,
+                                  builder: (context) {
+                                    return AlertDialog(
+                                      title: const Text('Delete Item'),
+                                      content: const Text(
+                                        'Are you sure you want to delete this item?',
+                                      ),
+                                      actions: [
+                                        TextButton(
+                                          onPressed: () {
+                                            Navigator.pop(
+                                              context,
+                                              false,
+                                            ); // User canceled
+                                          },
+                                          child: const Text('Cancel'),
+                                        ),
+                                        TextButton(
+                                          onPressed: () {
+                                            Navigator.pop(context, true);
+                                          },
+                                          child: const Text('Delete'),
+                                        ),
+                                      ],
+                                    );
+                                  },
+                                );
+
+                                if (confirmed == true) {
+                                  task.delete();
+                                  setState(() {});
+                                }
                               },
                             )
                           : null,
